+++ packages/orchestration/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dexto/orchestration
 
+## 1.6.25
+
+### Patch Changes
+
+- @dexto/core@1.6.25
+
 ## 1.6.24
 
 ### Patch Changes